Simplify matrix creation.
