+++ cpp/dolfin/la/utils.cpp
@@ -8,8 +8,14 @@
 #include <cassert>
 #include <dolfin/common/IndexMap.h>
 #include <dolfin/common/SubSystemsManager.h>
+#include <dolfin/la/SparsityPattern.h>
 #include <dolfin/log/log.h>
+#include <memory>
+
 #include <petsc.h>
+
+// Ceiling division of nonnegative integers
+#define dolfin_ceil_div(x, y) (x / y + int(x % y != 0))
 
 #define CHECK_ERROR(NAME)                                                      \
   do                                                                           \
@@ -66,6 +72,138 @@ Vec dolfin::la::create_vector(
   CHECK_ERROR("ISLocalToGlobalMappingDestroy");
 
   return x;
+}
+//-----------------------------------------------------------------------------
+Mat dolfin::la::create_matrix(
+    MPI_Comm comm, const dolfin::la::SparsityPattern& sparsity_pattern)
+{
+  PetscErrorCode ierr;
+  Mat A;
+  ierr = MatCreate(comm, &A);
+  if (ierr != 0)
+    petsc_error(ierr, __FILE__, "MatCreate");
+
+  // Get IndexMaps from sparsity patterm, and block size
+  std::array<std::shared_ptr<const common::IndexMap>, 2> index_maps
+      = {{sparsity_pattern.index_map(0), sparsity_pattern.index_map(1)}};
+  const int bs0 = index_maps[0]->block_size();
+  const int bs1 = index_maps[1]->block_size();
+
+  // Get global and local dimensions
+  const std::size_t M = bs0 * index_maps[0]->size_global();
+  const std::size_t N = bs1 * index_maps[1]->size_global();
+  const std::size_t m = bs0 * index_maps[0]->size_local();
+  const std::size_t n = bs1 * index_maps[1]->size_local();
+
+  // Find common block size across rows/columns
+  const int bs = (bs0 == bs1 ? bs0 : 1);
+  const int bs_map = bs;
+
+  // Set matrix size
+  ierr = MatSetSizes(A, m, n, M, N);
+  if (ierr != 0)
+    petsc_error(ierr, __FILE__, "MatSetSizes");
+
+  // Get number of nonzeros for each row from sparsity pattern
+  Eigen::Array<std::int32_t, Eigen::Dynamic, 1> nnz_diag
+      = sparsity_pattern.num_nonzeros_diagonal();
+  Eigen::Array<std::int32_t, Eigen::Dynamic, 1> nnz_offdiag
+      = sparsity_pattern.num_nonzeros_off_diagonal();
+
+  // Apply PETSc options from the options database to the matrix (this
+  // includes changing the matrix type to one specified by the user)
+  ierr = MatSetFromOptions(A);
+  if (ierr != 0)
+    petsc_error(ierr, __FILE__, "MatSetFromOptions");
+
+  // Build data to initialise sparsity pattern (modify for block size)
+  std::vector<PetscInt> _nnz_diag(nnz_diag.size() / bs),
+      _nnz_offdiag(nnz_offdiag.size() / bs);
+
+  for (std::size_t i = 0; i < _nnz_diag.size(); ++i)
+    _nnz_diag[i] = dolfin_ceil_div(nnz_diag[bs * i], bs);
+  for (std::size_t i = 0; i < _nnz_offdiag.size(); ++i)
+    _nnz_offdiag[i] = dolfin_ceil_div(nnz_offdiag[bs * i], bs);
+
+  // Allocate space for matrix
+  ierr = MatXAIJSetPreallocation(A, bs, _nnz_diag.data(), _nnz_offdiag.data(),
+                                 NULL, NULL);
+  if (ierr != 0)
+    petsc_error(ierr, __FILE__, "MatXIJSetPreallocation");
+
+  // Build and set local-to-global maps
+  assert(bs0 % bs_map == 0);
+  assert(bs1 % bs_map == 0);
+  std::vector<PetscInt> map0((m + index_maps[0]->num_ghosts())
+                             * (bs0 / bs_map));
+  std::vector<PetscInt> map1((n + index_maps[1]->num_ghosts())
+                             * (bs1 / bs_map));
+
+  const int row_block_size
+      = index_maps[0]->size_local() + index_maps[0]->num_ghosts();
+  for (int i = 0; i < row_block_size; ++i)
+  {
+    std::size_t factor = bs0 / bs_map;
+    auto index = index_maps[0]->local_to_global(i);
+    for (std::size_t j = 0; j < factor; ++j)
+      map0[i * factor + j] = factor * index + j;
+  }
+
+  const int col_block_size
+      = index_maps[1]->size_local() + index_maps[1]->num_ghosts();
+  for (int i = 0; i < col_block_size; ++i)
+  {
+    std::size_t factor = bs1 / bs_map;
+    auto index = index_maps[1]->local_to_global(i);
+    for (std::size_t j = 0; j < factor; ++j)
+      map1[i * factor + j] = factor * index + j;
+  }
+
+  // FIXME: In many cases the rows and columns could shared a common
+  // local-to-global map
+
+  // Create PETSc local-to-global map/index set
+  ISLocalToGlobalMapping petsc_local_to_global0, petsc_local_to_global1;
+  ierr = ISLocalToGlobalMappingCreate(MPI_COMM_SELF, bs_map, map0.size(),
+                                      map0.data(), PETSC_COPY_VALUES,
+                                      &petsc_local_to_global0);
+  if (ierr != 0)
+    petsc_error(ierr, __FILE__, "ISLocalToGlobalMappingCreate");
+  ierr = ISLocalToGlobalMappingCreate(MPI_COMM_SELF, bs_map, map1.size(),
+                                      map1.data(), PETSC_COPY_VALUES,
+                                      &petsc_local_to_global1);
+  if (ierr != 0)
+    petsc_error(ierr, __FILE__, "ISLocalToGlobalMappingCreate");
+
+  // Set matrix local-to-global maps
+  ierr = MatSetLocalToGlobalMapping(A, petsc_local_to_global0,
+                                    petsc_local_to_global1);
+  if (ierr != 0)
+    petsc_error(ierr, __FILE__, "MatSetLocalToGlobalMappingXXX");
+
+  // Note: This should be called after having set the local-to-global
+  // map for MATIS (this is a dummy call if A is not of type MATIS)
+  ierr = MatISSetPreallocation(A, 0, _nnz_diag.data(), 0, _nnz_offdiag.data());
+  if (ierr != 0)
+    petsc_error(ierr, __FILE__, "MatISSetPreallocation");
+
+  // Clean up local-to-global maps
+  ierr = ISLocalToGlobalMappingDestroy(&petsc_local_to_global0);
+  if (ierr != 0)
+    petsc_error(ierr, __FILE__, "ISLocalToGlobalMappingDestroy");
+  ierr = ISLocalToGlobalMappingDestroy(&petsc_local_to_global1);
+  if (ierr != 0)
+    petsc_error(ierr, __FILE__, "ISLocalToGlobalMappingDestroy");
+
+  // Set some options on Mat object
+  ierr = MatSetOption(A, MAT_NEW_NONZERO_ALLOCATION_ERR, PETSC_TRUE);
+  if (ierr != 0)
+    petsc_error(ierr, __FILE__, "MatSetOption");
+  ierr = MatSetOption(A, MAT_KEEP_NONZERO_PATTERN, PETSC_TRUE);
+  if (ierr != 0)
+    petsc_error(ierr, __FILE__, "MatSetOption");
+
+  return A;
 }
 //-----------------------------------------------------------------------------
 std::vector<IS> dolfin::la::compute_index_sets(
